chore: Fixed links to OSS friends (#1006)
The links on
[https://docs.tolgee.io/oss-friends](https://docs.tolgee.io/oss-friends)
were leading to broken pages with the URL like:
https://docs.tolgee.io/https://activepieces.com

Now it opens the websites properly

P.S. I am not sure if the "rel" attribute in the links should be what it
is, as I am not aware of what conditions the links are kept there in the
docs. Maybe they need to be dofollow?